+++ src/jquery.taboverride.js
@@ -1,4 +1,4 @@
-/*global define, jQuery, TABOVERRIDE */
+/*global exports, require, define, jQuery, TABOVERRIDE */
 
 /**
  * the global jQuery object
@@ -15,17 +15,20 @@
  * @memberOf jQuery
  */
 
-// Use AMD if available
+// Use CommonJS or AMD if available
 (function ( factory ) {
     "use strict";
 
-    if ( typeof define === "function" && define.amd ) {
-        // AMD - Register as an anonymous module
-        // Files must be concatenated using an AMD-aware tool such as r.js
-        define( [ "jquery", "taboverride" ], factory );
+    if ( typeof exports === "object" ) {
+	// Node/CommonJS
+	factory( require( "jquery", "taboverride" ) );
+    } else if ( typeof define === "function" && define.amd ) {
+	// AMD - Register as an anonymous module
+	// Files must be concatenated using an AMD-aware tool such as r.js
+	define( [ "jquery", "taboverride" ], factory );
     } else {
-        // No AMD - Use global variables instead
-        factory( jQuery, TABOVERRIDE );
+	// No AMD - Use global variables instead
+	factory( jQuery, TABOVERRIDE );
     }
 }(function ( $, TABOVERRIDE ) {
     "use strict";
@@ -45,10 +48,10 @@
      */
     var tabOverride = $.fn.tabOverride = function ( enable, selector ) {
 
-        var enablePlugin = !arguments.length || enable,
+	var enablePlugin = !arguments.length || enable,
             isDelegated = typeof selector === "string";
 
-        if ( isDelegated ) {
+	if ( isDelegated ) {
             // Always remove the event handlers
             this.off( ".tabOverride", selector );
 
@@ -61,10 +64,10 @@
             // The jQuery object acts as an array of elements, so it can be passed
             // to TABOVERRIDE.set()
             TABOVERRIDE.set( this, enablePlugin );
-        }
+	}
 
-        // Return the jQuery object
-        return this;
+	// Return the jQuery object
+	return this;
     };
 
     /**