+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Updated build process
   * Use JSHint for all js files
   * Generate API documentation with JSDoc 3
+* Support CommonJS
 
 ## 3.2.3 (2012-01-21)
 * Improved description and dependencies in manifest file